requirements.txt
fastapi
uvicorn
jupyterlab
n8n
pillow
pydantic
numpy
scipy
# ggf. für Text2Voice:
elevenlabs
# ggf. für Musik / Bild-Modelle:
torch
diffusers
transformers
